deployment/requirements.txt
flask==3.0.3
mlflow==2.17.0
python-dotenv==0.19.0
requests==2.32.2